+++ src/JavaScriptRuntime/GlobalThis.cs
@@ -287,6 +287,9 @@ namespace JavaScriptRuntime
         private static readonly Func<object[], object?[], object?> _uriErrorConstructorValue =
             CreateErrorConstructorValue(static message => new JavaScriptRuntime.URIError(message));
 
+        private static readonly Func<object[], object?[], object?> _aggregateErrorConstructorValue = static (_, args) =>
+            JavaScriptRuntime.AggregateError.Construct(args ?? System.Array.Empty<object?>());
+
         private static readonly Func<object[], object?[], object?> _iteratorConstructorValue = static (_, __) =>
             throw new TypeError("Iterator is not directly constructible in jroc.");
 
@@ -304,6 +307,7 @@ namespace JavaScriptRuntime
         private static readonly object _syntaxErrorPrototypeValue = new JsObject();
         private static readonly object _typeErrorPrototypeValue = new JsObject();
         private static readonly object _uriErrorPrototypeValue = new JsObject();
+        private static readonly object _aggregateErrorPrototypeValue = new JsObject();
 
         // Minimal Object.prototype object used for descriptor/prototype-heavy libraries.
         // NOTE: We intentionally do not enable PrototypeChain here; Object.create/setPrototypeOf
@@ -684,6 +688,7 @@ namespace JavaScriptRuntime
             ConfigureErrorIntrinsicSurface(_syntaxErrorConstructorValue, _syntaxErrorPrototypeValue, "SyntaxError", parentPrototype: _errorPrototypeValue);
             ConfigureErrorIntrinsicSurface(_typeErrorConstructorValue, _typeErrorPrototypeValue, "TypeError", parentPrototype: _errorPrototypeValue);
             ConfigureErrorIntrinsicSurface(_uriErrorConstructorValue, _uriErrorPrototypeValue, "URIError", parentPrototype: _errorPrototypeValue);
+            ConfigureAggregateErrorIntrinsicSurface();
 
             PropertyDescriptorStore.DefineOrUpdate(_booleanPrototypeValue, "constructor", new JsPropertyDescriptor
             {
@@ -1371,6 +1376,9 @@ namespace JavaScriptRuntime
             dict.TryAdd(nameof(GlobalThis.URIError), URIError);
             DefineNonEnumerableDataProperty(nameof(GlobalThis.URIError), dict[nameof(GlobalThis.URIError)]);
 
+            dict.TryAdd(nameof(GlobalThis.AggregateError), AggregateError);
+            DefineNonEnumerableDataProperty(nameof(GlobalThis.AggregateError), dict[nameof(GlobalThis.AggregateError)]);
+
             dict.TryAdd(nameof(GlobalThis.Iterator), Iterator);
             DefineNonEnumerableDataProperty(nameof(GlobalThis.Iterator), dict[nameof(GlobalThis.Iterator)]);
 
@@ -1658,6 +1666,8 @@ namespace JavaScriptRuntime
         public static Func<object[], object?[], object?> TypeError => _typeErrorConstructorValue;
 
         public static Func<object[], object?[], object?> URIError => _uriErrorConstructorValue;
+
+        public static Func<object[], object?[], object?> AggregateError => _aggregateErrorConstructorValue;
 
         public static Func<object[], object?[], object?> Iterator => _iteratorConstructorValue;
 
@@ -2296,6 +2306,7 @@ namespace JavaScriptRuntime
         internal static object SyntaxErrorPrototypeValue => _syntaxErrorPrototypeValue;
         internal static object TypeErrorPrototypeValue => _typeErrorPrototypeValue;
         internal static object URIErrorPrototypeValue => _uriErrorPrototypeValue;
+        internal static object AggregateErrorPrototypeValue => _aggregateErrorPrototypeValue;
         private static Func<object[], object?[], object?> CreateErrorConstructorValue(Func<string?, object> factory)
         {
             return (_, args) =>
@@ -2358,6 +2369,33 @@ namespace JavaScriptRuntime
             });
         }
 
+        private static void ConfigureAggregateErrorIntrinsicSurface()
+        {
+            ConfigureErrorIntrinsicSurface(
+                _aggregateErrorConstructorValue,
+                _aggregateErrorPrototypeValue,
+                "AggregateError",
+                _errorPrototypeValue);
+            PrototypeChain.SetPrototype(_aggregateErrorConstructorValue, _errorConstructorValue);
+
+            PropertyDescriptorStore.DefineOrUpdate(_aggregateErrorConstructorValue, "length", new JsPropertyDescriptor
+            {
+                Kind = JsPropertyDescriptorKind.Data,
+                Enumerable = false,
+                Configurable = true,
+                Writable = false,
+                Value = 2d
+            });
+            PropertyDescriptorStore.DefineOrUpdate(_aggregateErrorConstructorValue, "prototype", new JsPropertyDescriptor
+            {
+                Kind = JsPropertyDescriptorKind.Data,
+                Enumerable = false,
+                Configurable = false,
+                Writable = false,
+                Value = _aggregateErrorPrototypeValue
+            });
+        }
+
         internal static void AssignBuiltInErrorPrototype(JavaScriptRuntime.Error error)
         {
             ArgumentNullException.ThrowIfNull(error);
@@ -2371,6 +2409,7 @@ namespace JavaScriptRuntime
                 JavaScriptRuntime.SyntaxError => _syntaxErrorPrototypeValue,
                 JavaScriptRuntime.TypeError => _typeErrorPrototypeValue,
                 JavaScriptRuntime.URIError => _uriErrorPrototypeValue,
+                JavaScriptRuntime.AggregateError => _aggregateErrorPrototypeValue,
                 _ => _errorPrototypeValue
             };